Before leaving for Christmas/Holiday vacation, some changes
Updated the documentation links to admin+api+website to use localhost and removed the port. Ref #62

Updated the documentation: Using a single line to run manage.py commands. Ref #62

Moved the import_sk3/migrate_sk3 script so that it can be run as a custom manage.py command. (And documented this). Ref #51