WIP
Updated models.py with the new Region (and link to Initiative), and updated serializers.py, views.py, urls.py, admin.py to match. Added migrations

Updated import script to import the region (and link to initiative)

Added directory docs/sk3_api_responses and added 2x2 files there for regions and for gbg business. This is important when working with the import of data

Some documentation housekeeping